requirements.txt
streamlit
asyncio
langchain_google_genai
browser_use
google-cloud-bigquery
google-auth
streamlit-feedback
requests
pandas
python-dotenv